# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Planiranje Putovanja API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/aktivnosti/search": {
            "get": {
                "tags": [
                    "Aktivnosti"
                ],
                "summary": "Pretraga aktivnosti po nazivu ili opisu",
                "description": "Omogućava pretragu svih dostupnih aktivnosti na osnovu unetog teksta.",
                "operationId": "eb60a2f2686dc1eec701a969ded94b53",
                "parameters": [
                    {
                        "name": "query",
                        "in": "query",
                        "description": "Tekst za pretragu",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Uspešna pretraga",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "naziv": {
                                                "type": "string",
                                                "example": "Obilazak Luvra"
                                            },
                                            "opis": {
                                                "type": "string",
                                                "example": "Detaljan obilazak muzeja sa vodičem."
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Nije pronađena nijedna aktivnost sa tim pojmom"
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Autentifikacija"
                ],
                "summary": "Registracija novog korisnika",
                "operationId": "8a56853624e025573120a09a4c75d468",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Korisnik uspešno kreiran"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Autentifikacija"
                ],
                "summary": "Logovanje korisnika",
                "description": "Korisnik šalje kredencijale da bi dobio Bearer token.",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "admin@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Uspešan login",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string"
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "Bearer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Pogrešan email ili lozinka"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Autentifikacija"
                ],
                "summary": "Odjava korisnika",
                "description": "Uništava trenutni pristupni token i odjavljuje korisnika sa sistema.",
                "operationId": "fe8f3429cd6979b3b4517e186505f9f9",
                "responses": {
                    "200": {
                        "description": "Uspešna odjava",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Logged out successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Niste autorizovani (nedostaje ili je neispravan token)"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/users": {
            "get": {
                "tags": [
                    "Autentifikacija"
                ],
                "summary": "Prikaz svih korisnika",
                "operationId": "ce757d6a9dd9f3bf15fabb6611112518",
                "responses": {
                    "200": {
                        "description": "Lista korisnika"
                    },
                    "403": {
                        "description": "Samo admin može videti listu"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/users/{id}": {
            "delete": {
                "tags": [
                    "Autentifikacija"
                ],
                "summary": "Brisanje korisnika",
                "operationId": "8d623c822968776003e4d1009d1321b8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Korisnik obrisan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}